Byte Size: Specify long instead of int32

+++ src/main/java/org/damap/base/rda/dmpcommonstandard/DatasetMapper.java
@@ -69,7 +69,7 @@ public class DatasetMapper extends AbstractMapper {
     }
     var size = datasetDO.getSize();
     if (size != null) {
-      distribution.setByteSize(Math.toIntExact(size));
+      distribution.setByteSize(size);
     }
     result.setDescription(datasetDO.getDescription());
     var personalData = datasetDO.getPersonalData();
